backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
Pillow==10.0.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
drf-nested-routers==0.93.4
django-filter==23.3
redis==5.0.1
PyMySQL==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
